requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pandas==2.1.3
joblib==1.3.2
python-multipart==0.0.6
requests==2.31.0
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1
streamlit==1.28.1
plotly==5.17.0
numpy==1.24.3
python-dotenv==1.0.0
scikit-learn==1.3.2
xgboost==1.7.6
clearml==1.13.2